refactor parser state out
