+++ lib/parser.js
@@ -6,27 +6,8 @@ var classHandlers = require('./handlers/classHandlers');
 var functionHandlers = require('./handlers/functionHandlers');
 var moduleHandlers = require('./handlers/moduleHandlers');
 
-var multiComment = false;
-var multiLine = false;
-
-var stateStack = []
-
-var StateEnum = {
-  CONTROL_FLOW: 1
-}
-
-/**
- * Determines whether the parser has entered the desired state or not
- */
-function hasState(state)
-{
-  if (stateStack.length < 1)
-  {
-    return false;
-  }
-
-  return stateStack[stateStack.length-1] == state;
-}
+var parserState = require('./parserState');
+const StateEnum = parserState.StateEnum
 
 var validOperators = {
     'is': ' === ',
@@ -168,13 +149,13 @@ function isDogescriptSource(parseContext)
   }
 
   // ending a multi comment block
-  if (multiComment && tokens[0] === 'loud' )
+  if (parserState.hasState(StateEnum.MULTILINE_COMMENT) && tokens[0] === 'loud' )
   {
     return true;
   }
 
   // closing a multi-line object creation
-  if ( multiLine && tokens[0] === 'wow' )
+  if ( parserState.hasState(StateEnum.OBJECT) && tokens[0] === 'wow' )
   {
     return true;
   }
@@ -239,7 +220,7 @@ function shouldCloseStatement(parseContext, statement)
   }
 
   // if we're in multiline mode, the property:value assignments should not have a ';'
-  return !multiLine;
+  return !parserState.hasState(StateEnum.OBJECT);
 }
 
 /**
@@ -312,7 +293,7 @@ function handleShh(parseContext)
    // consume: quiet
    tokens.shift();
 
-   multiComment = true;
+   parserState.pushState(StateEnum.MULTILINE_COMMENT);
 
    var statement = '/*';
    statement += tokenUtils.joinTokens(parseContext.tokens);
@@ -333,7 +314,7 @@ function handleLoud(parseContext)
 
   var tokens = parseContext.tokens;
 
-  if(!multiComment)
+  if(!parserState.hasState(StateEnum.MULTILINE_COMMENT))
   {
     throw new Error("Unparseable syntax! Encountered: 'loud' without first seeing 'quiet'");
   }
@@ -341,7 +322,8 @@ function handleLoud(parseContext)
   // consume: loud
   tokens.shift();
 
-  multiComment = false;
+  // Remove multicomment state
+  parserState.popState();
 
   var statement = '*/';
   statement += tokenUtils.joinTokens(parseContext.tokens);
@@ -403,9 +385,9 @@ function handleRly(parseContext)
   tokens.shift();
 
   var statement = 'if ('
-  stateStack.push(StateEnum.CONTROL_FLOW);
+  parserState.pushState(StateEnum.CONTROL_FLOW);
   statement += parseStatements(parseContext);
-  stateStack.pop();
+  parserState.popState();
   // close condition and open branch
   statement += ') {\n';
   return statement;
@@ -425,9 +407,9 @@ function handleNotrly(parseContext)
   tokens.shift();
 
   var statement = 'if (!';
-  stateStack.push(StateEnum.CONTROL_FLOW);
+  parserState.pushState(StateEnum.CONTROL_FLOW);
   statement += parseStatements(parseContext);
-  stateStack.pop();
+  parserState.popState();
   // close condition and open branch
   statement += ') {\n';
   return statement;
@@ -573,7 +555,11 @@ function handleWow(parseContext)
   tokenUtils.expectAnyToken(['wow', 'wow&'], parseContext);
 
   var tokens = parseContext.tokens;
-  multiLine = false;
+  // turn off multiline declaration
+  if(parserState.hasState(StateEnum.OBJECT))
+  {
+    parserState.popState();
+  }
 
   var chained = tokens[0] === 'wow&';
 
@@ -624,9 +610,9 @@ function handleMuchLoop(parseContext)
   tokens.shift();
 
   var statement = 'for ('
-  stateStack.push(StateEnum.CONTROL_FLOW);
+  parserState.pushState(StateEnum.CONTROL_FLOW);
   statement += parseStatements(parseContext);
-  stateStack.pop();
+  parserState.popState();
   statement += ') {\n';
   return statement;
 }
@@ -648,9 +634,9 @@ function handleMany(parseContext)
   tokens.shift();
 
   var statement = 'while (';
-  stateStack.push(StateEnum.CONTROL_FLOW);
+  parserState.pushState(StateEnum.CONTROL_FLOW);
   statement += parseStatements(parseContext);
-  stateStack.pop();
+  parserState.popState();
   statement += ') {\n';
 
   return statement;
@@ -841,7 +827,7 @@ function handleVery(parseContext) {
     // consume obj
     tokens.shift();
     statement += '{\n';
-    multiLine = true;
+    parserState.pushState(StateEnum.OBJECT);
     return statement;
   }
 
@@ -952,7 +938,7 @@ function parseStatement(parseContext) {
   var tokens = parseContext.tokens;
 
   // if processing multi-comment
-  if(multiComment)
+  if(parserState.hasState(StateEnum.MULTILINE_COMMENT))
   {
     // consume tokens until we encounter loud
     if(tokens[0] !== 'loud')
@@ -1005,7 +991,7 @@ function parseStatement(parseContext) {
   }
 
   // only applicable to disambiguate during control flow parsing
-  if ( hasState(StateEnum.CONTROL_FLOW) )
+  if ( parserState.hasState(StateEnum.CONTROL_FLOW) )
   {
     // convert to supported operator / otherwise fall thorugh
     // TODO remove operators as they become supported
@@ -1110,7 +1096,7 @@ function parseStatement(parseContext) {
   if (tokens[0] === 'obj') {
     // consume obj
     tokens.shift();
-    multiLine = true;
+    parserState.pushState(StateEnum.OBJECT);
     return '{\n';
   }
